// ui/src/generated/models/list-albums-response.ts
/* tslint:disable */
/* eslint-disable */
/**
 * Photos NG API
 * API for managing albums and media in Photos NG application
 *
 * The version of the OpenAPI document: 1.0.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


// May contain unused imports in some cases
// @ts-ignore
import type { Album } from './album';

/**
 * 
 * @export
 * @interface ListAlbumsResponse
 */
export interface ListAlbumsResponse {
    /**
     * 
     * @type {Array<Album>}
     * @memberof ListAlbumsResponse
     */
    'albums': Array<Album>;
    /**
     * Total number of albums
     * @type {number}
     * @memberof ListAlbumsResponse
     */
    'total': number;
    /**
     * Number of albums returned
     * @type {number}
     * @memberof ListAlbumsResponse
     */
    'limit': number;
    /**
     * Number of albums skipped
     * @type {number}
     * @memberof ListAlbumsResponse
     */
    'offset': number;
}